Use int32_t and uint32_t in unistd syscalls

+++ libc/unistd/getticks.c
@@ -1,8 +1,9 @@
+#include <stdint.h>
 #include <sys/syscall.h>
 #include <unistd.h>
 
 unsigned long long getticks(void) {
-  unsigned int lo, hi;
+  uint32_t lo, hi;
   asm volatile("int $0x80" : "=a"(lo), "=d"(hi) : "a"(SYS_GETTICKS));
-  return ((unsigned long long)hi << 32) | lo;
+  return ((uint64_t)hi << 32) | lo;
 }
